backend/requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.3
Werkzeug==2.2.2
zipp==3.11.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paypalrestsdk==1.13.1
pycparser==2.21
pymongo==4.3.3
pyOpenSSL==22.1.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
